- Убран экран статистики - Теперь выбирать диапазон дат можно на главном экране - Пайчарт будет отображаться при выборе диапазона дат вместо недельных графиков - Автофокус текстового поля названия транзакции при создании транзакции - Автофокус текстового поля названия категории при создании категории - Новое отображение инпута выбора категории

+++ lib/presentation/ui/main/components/app_bar.dart
@@ -1,8 +1,14 @@
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:my_expenses_planner/config/constants.dart';
 import 'package:my_expenses_planner/config/l10n/localization.dart';
+import 'package:my_expenses_planner/core/extensions/date_time_extensions.dart';
+import 'package:my_expenses_planner/core/extensions/date_time_range_extensions.dart';
 import 'package:my_expenses_planner/di.dart';
 import 'package:my_expenses_planner/domain/use_cases/transactions/i_transactions_case.dart';
+import 'package:my_expenses_planner/presentation/cubit/transaction_list/transaction_list_cubit.dart';
+import 'package:my_expenses_planner/presentation/ui/core/widgets/date_range_picker.dart';
 
 class MainScreenAppBar extends StatelessWidget implements PreferredSizeWidget {
   const MainScreenAppBar({Key? key}) : super(key: key);
@@ -12,24 +18,58 @@ class MainScreenAppBar extends StatelessWidget implements PreferredSizeWidget {
 
   @override
   Widget build(BuildContext context) {
-    return AppBar(
-      title: Text(AppLocalizationsWrapper.of(context).main_app_bar_title),
-      actions: !ConfigConstants.isTest
-          ? null
-          : [
+    return BlocBuilder<TransactionListCubit, TransactionListState>(
+      builder: (context, state) {
+        return AppBar(
+          centerTitle: true,
+          title: AutoSizeText(
+            state.dateTimeRange == null
+                ? AppLocalizationsWrapper.of(context).main_app_bar_title
+                : state.dateTimeRange!.toFormattedString(context),
+            maxLines: 1,
+          ),
+          actions: [
+            if (!ConfigConstants.isTest)
               IconButton(
                 onPressed: () {
                   getIt<ITransactionsCase>().deleteAll();
                 },
                 icon: const Icon(Icons.delete),
               ),
+            if (!ConfigConstants.isTest)
               IconButton(
                 onPressed: () {
                   getIt<ITransactionsCase>().fillWithMockTransactions();
                 },
                 icon: const Icon(Icons.add),
               ),
-            ],
+            CustomDateRangePickerIcon(
+              initialDateTimeRange: state.dateTimeRange,
+              onDateTimeRangePicked: (DateTimeRange? dateTimeRange) {
+                if (dateTimeRange != null) {
+                  final _dateTimeRange = DateTimeRange(
+                    start: dateTimeRange.start.startOfDay,
+                    end: dateTimeRange.end.endOfDay,
+                  );
+
+                  BlocProvider.of<TransactionListCubit>(context)
+                      .onDateTimeRangeChange(_dateTimeRange);
+                }
+              },
+            ),
+            if (state.dateTimeRange != null)
+              IconButton(
+                onPressed: () {
+                  BlocProvider.of<TransactionListCubit>(context)
+                      .onDateTimeRangeChange(null);
+                },
+                icon: const Icon(
+                  Icons.close,
+                ),
+              ),
+          ],
+        );
+      },
     );
   }
 }
